+++ bcbio/cwl/workflow.py
@@ -269,6 +269,14 @@ def _handle_special_inputs(inputs, variables):
                     out.append(vid)
                     found_indexes = True
             assert found_indexes, "Found no aligner indexes in %s" % [v["id"] for v in variables]
+        elif input == ["reference", "snpeff", "genome_build"]:
+            found_indexes = False
+            for v in variables:
+                vid = get_base_id(v["id"]).split("__")
+                if vid[0] == "reference" and vid[1] == "snpeff":
+                    out.append(vid)
+                    found_indexes = True
+            assert found_indexes, "Found no snpEff indexes in %s" % [v["id"] for v in variables]
         elif input in optional:
             if _get_string_vid(input) in all_vs:
                 out.append(input)